feat: split border stories
