+++ packages/storybook/src/primitive-tokens/BorderWidth.stories.tsx
@@ -7,23 +7,16 @@ import { path2css } from '../denhaag/util';
 type Story = StoryObj<typeof meta>;
 type Tokens = typeof tokens;
 type BorderWidthTokens = Tokens['denhaag']['border-width'];
-type BorderStyleTokens = Tokens['denhaag']['border-style'];
 
 const borderWidthTokens: BorderWidthTokens = tokens['denhaag']['border-width'];
-const borderStyleTokens: BorderStyleTokens = tokens['denhaag']['border-style'];
 const borderWidthTokensSorted = Object.entries(borderWidthTokens).sort((a, b) => {
-  const aValue = parseFloat(a[1].value);
-  const bValue = parseFloat(b[1].value);
-  return aValue - bValue;
-});
-const borderStyleTokensSorted = Object.entries(borderStyleTokens).sort((a, b) => {
   const aValue = parseFloat(a[1].value);
   const bValue = parseFloat(b[1].value);
   return aValue - bValue;
 });
 
 const meta: Meta = {
-  title: 'Primitive Tokens/Border',
+  title: 'Primitive Tokens/Border Width',
   tags: ['autodocs', '!dev'],
   parameters: {
     chromatic: { disableSnapshot: true },
@@ -57,38 +50,6 @@ const meta: Meta = {
                         width: '40px',
                         borderStyle: 'solid',
                         borderWidth: width.value,
-                      }}
-                    ></div>
-                  </td>
-                </tr>
-              ))}
-            </tbody>
-          </table>
-          <Subtitle>Style</Subtitle>
-          <table>
-            <thead>
-              <tr>
-                <th>Name</th>
-                <th>Variable</th>
-                <th>Value</th>
-                <th>Preview</th>
-              </tr>
-            </thead>
-            <tbody>
-              {borderStyleTokensSorted.map(([key, style]) => (
-                <tr key={key}>
-                  <td>{key}</td>
-                  <td>
-                    <code>{path2css(style.path)}</code>
-                  </td>
-                  <td>{style.value}</td>
-                  <td>
-                    <div
-                      style={{
-                        margin: '0 auto',
-                        height: '40px',
-                        width: '40px',
-                        borderStyle: style.value,
                       }}
                     ></div>
                   </td>
